refactor(screensaver): extract ScreenSaverAssetLoader from ScreenSaverService
Move the screensaver's asset loading and MediaSliderConfiguration assembly out
of the DreamService into a Context-agnostic ScreenSaverAssetLoader, so a second
host can drive the exact same pipeline without duplicating it.

The service keeps its behaviour: it now implements ScreenSaverAssetLoader.Host
and supplies the three UI side effects (toast, finish, load-into-view). Prefs
are still read at their original call sites, and the existing
maxCutOffWidth = SLIDER_MAX_CUT_OFF_HEIGHT quirk is preserved verbatim.

One deliberate delta: the album error paths previously called finish() from the
IO thread while only the toast hopped to Main. Both now run on Main, which is
safer for the dream and required for a fragment host.

Add ScreenSaverAssetFilterTest covering the extracted asset filtering.
